login page something update

+++ src/Pages/MyWatchListTable/MyWatchListTable.jsx
@@ -6,21 +6,38 @@ const MyWatchListTable = ({ listData, watchListData, setWatchListData }) => {
   const { _id, photo, title, genres, newEmail, rating } = listData;
 
   const handleWatchList = (id) => {
-    fetch(`https://game-review-server-umber.vercel.app/watchList/${id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        if (data.deletedCount > 0) {
-          Swal.fire({
-            title: "Deleted!",
-            text: "Deleted from your watchList.",
-            icon: "success",
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(`https://game-review-server-umber.vercel.app/watchList/${id}`, {
+          method: "DELETE",
+        })
+          .then((res) => res.json())
+          .then((data) => {
+            if (data.deletedCount > 0) {
+              Swal.fire({
+                title: "Deleted!",
+                text: "Deleted from your watchList.",
+                icon: "success",
+              });
+              const remainingWatchList = watchListData.filter(fil => fil._id !== id)
+              setWatchListData(remainingWatchList)
+            }
           });
-          const remainingWatchList = watchListData.filter(fil => fil._id !== id)
-          setWatchListData(remainingWatchList)
-        }
-      });
+      }
+    });
+
+
+
+
+   
   };
   return (
     <>
